Extract WMS config and longpress modules, improve toolbar UX and accessibility
Phase 0 extraction:
- Extract WMS server configuration to src/config.js (~138 lines out of radar.js)
- Extract createLongPressHandler to src/longpress.js with decoupled callback API

Toolbar UX improvements:
- Change toolbar buttons from <div> to semantic <button> elements
- Change toolbar container from <div> to <nav>
- Add ARIA attributes: aria-pressed, aria-label, aria-haspopup
- Add stronger selected state with background highlight and accent border
- Add long-press affordance indicator (expand_more arrow) on menu buttons
- Add press feedback (scale + highlight) via JS-driven .pressing class
- Add :focus-visible ring for keyboard navigation
- Add CSS transitions (150ms) for smoother state changes
- Ensure equal button heights with invisible spacers
- Set min-height 48px for touch target compliance
- Consolidate setButtonStates() using new setButtonState() helper
